third_party/blink/web_tests/fast/xsl/xslt/xslt-translate.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<script>
description("Must not crash, and transforms to 'Simple_sting_line'. Test for http://crbug.com/599427.");

var stylesheetXml =
    '<xsl:stylesheet xmlns:xsl="http://www.w3.org/1999/XSL/Transform" version="1.0">' +
        '<xsl:template match="/">' +
            '<div>' +
                '<xsl:value-of select=\'translate("Simple sting line"," ","_")\'/>' +
            '</div>' +
        '</xsl:template>' +
    '</xsl:stylesheet>';

var parser = new DOMParser();
var doc = parser.parseFromString(stylesheetXml, 'text/xml');
var xslt = parser.parseFromString(stylesheetXml, 'text/xml');
var processor = new XSLTProcessor();
processor.importStylesheet(xslt);

shouldBe('processor.transformToFragment(doc, document).textContent', '"Simple_sting_line"');
</script>
